+++ bee-common/bee-common/CHANGELOG.md
@@ -26,7 +26,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ord::is_unique_sorted`;
 - `ord::is_sorted`;
 - `impl<const N: usize> Packable for [u8; N]`;
-- `Packable::unpack_inner<R: Read + ?Sized, const TRUSTED: bool>`;
+- `Packable::unpack_inner`;
+- `Packable::unpack_unchecked`;
 
 ### Changed