+++ bee-common/bee-common/CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security -->
 
-## 0.4.0 - 2021-04-07
+## 0.4.0 - 2021-04-08
 
 ### Added